+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # CHANGELOG
 
 ## 1.3 (unreleased)
-- `.as_email_message` -> EmailMessage
+- `.as_message` -> Message
+- e-mail validation
+- proper e-mail addresses handling: they can be in a string delimited by comma (or semicolon)
 
 ## 1.2 (2020-06-18)
 - fix: smtplib.SMTP_SSL handshake operation timed out caught